requirements.txt
psycopg2-binary
pytest
pytest-mock
python-dotenv


